+++ crates/objc2/src/__framework_prelude.rs
@@ -14,8 +14,8 @@ pub use std::os::raw::{
 pub use crate::encode::{Encode, Encoding, RefEncode};
 pub use crate::ffi::{NSInteger, NSIntegerMax, NSUInteger, NSUIntegerMax};
 pub use crate::mutability::{
-    InteriorMutable, InteriorMutableWithSubclass, InteriorMutableWithSuperclass, IsIdCloneable,
-    IsMainThreadOnly, IsRetainable, MainThreadOnly,
+    InteriorMutable, InteriorMutableWithSubclass, InteriorMutableWithSuperclass, IsMainThreadOnly,
+    MainThreadOnly,
 };
 pub use crate::rc::{Allocated, DefaultId, DefaultRetained, Id, Retained};
 pub use crate::runtime::{