+++ crates/test-ui/ui/mutability_traits_unimplementable2.rs
@@ -3,8 +3,6 @@ use objc2::mutability;
 
 struct CustomStruct;
 
-unsafe impl mutability::IsIdCloneable for CustomStruct {}
-
 unsafe impl mutability::IsAllocableAnyThread for CustomStruct {}
 
 fn main() {}