+++ framework-crates/objc2-foundation/src/to_owned.rs
@@ -1,14 +1,13 @@
 #![allow(unused_imports)]
 use alloc::borrow::ToOwned;
 
-use objc2::mutability::IsIdCloneable;
 use objc2::rc::Retained;
 use objc2::Message;
 
 use crate::Foundation::{self, NSCopying, NSMutableCopying};
 
 #[cfg(feature = "NSArray")]
-impl<T: Message + IsIdCloneable> ToOwned for Foundation::NSArray<T> {
+impl<T: Message> ToOwned for Foundation::NSArray<T> {
     type Owned = Retained<Self>;
     fn to_owned(&self) -> Self::Owned {
         self.copy()
@@ -16,7 +15,7 @@ impl<T: Message + IsIdCloneable> ToOwned for Foundation::NSArray<T> {
 }
 
 #[cfg(feature = "NSArray")]
-impl<T: Message + IsIdCloneable> ToOwned for Foundation::NSMutableArray<T> {
+impl<T: Message> ToOwned for Foundation::NSMutableArray<T> {
     type Owned = Retained<Self>;
     fn to_owned(&self) -> Self::Owned {
         self.mutableCopy()
@@ -48,7 +47,7 @@ impl ToOwned for Foundation::NSException {
 }
 
 #[cfg(feature = "NSSet")]
-impl<T: Message + IsIdCloneable> ToOwned for Foundation::NSSet<T> {
+impl<T: Message> ToOwned for Foundation::NSSet<T> {
     type Owned = Retained<Self>;
     fn to_owned(&self) -> Self::Owned {
         self.copy()
@@ -56,7 +55,7 @@ impl<T: Message + IsIdCloneable> ToOwned for Foundation::NSSet<T> {
 }
 
 #[cfg(feature = "NSSet")]
-impl<T: Message + IsIdCloneable> ToOwned for Foundation::NSMutableSet<T> {
+impl<T: Message> ToOwned for Foundation::NSMutableSet<T> {
     type Owned = Retained<Self>;
     fn to_owned(&self) -> Self::Owned {
         self.mutableCopy()