+++ framework-crates/objc2-foundation/src/util.rs
@@ -2,7 +2,6 @@
 use core::ptr::NonNull;
 
 use objc2::rc::Retained;
-use objc2::Message;
 
 pub(crate) fn retained_ptr_cast<T: ?Sized>(objects: *mut Retained<T>) -> *mut NonNull<T> {
     // SAFETY: `Retained<T>` has the same memory layout as `NonNull<T>`, and
@@ -18,13 +17,4 @@ pub(crate) fn ref_ptr_cast_const<T: ?Sized>(objects: *const &T) -> *mut NonNull<
 
 pub(crate) fn retained_ptr_cast_const<T: ?Sized>(objects: *const Retained<T>) -> *mut NonNull<T> {
     retained_ptr_cast(objects as *mut Retained<T>)
-}
-
-/// # Safety
-///
-/// Currently unsound, but will be sound after changes to `objc2`.
-#[inline]
-pub(crate) fn retain<T: Message>(obj: &T) -> Retained<T> {
-    // SAFETY: TODO
-    unsafe { Retained::retain(obj as *const T as *mut T).unwrap_unchecked() }
 }